# perf(core): only devirtualize when ident matches a workspace