**What's the problem this PR addresses?**

`Project.tryWorkspaceByDescriptor` and `Project.tryWorkspaceByLocator` devirtualizes the descriptor/locator even if the ident doesn't match a workspace.

**How did you fix it?**

Check if the descriptor/locator matches a workspace before devirtualizing them

**Result**

Tested on the Storybook repo using PnP
```
$ YARN_IGNORE_PATH=1 hyperfine -w 5 'nodejs before.js' 'nodejs after.js'
Benchmark #1: nodejs before.js
  Time (mean ± σ):      5.217 s ±  0.047 s    [User: 5.538 s, System: 1.219 s]
  Range (min … max):    5.179 s …  5.339 s    10 runs

Benchmark #2: nodejs after.js
  Time (mean ± σ):      5.156 s ±  0.013 s    [User: 5.535 s, System: 1.194 s]
  Range (min … max):    5.137 s …  5.176 s    10 runs

Summary
  'nodejs after.js' ran
    1.01 ± 0.01 times faster than 'nodejs before.js'
```

**Checklist**
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).
- [x] I have set the packages that need to be released for my changes to be effective.
- [x] I will check that all automated PR checks pass before the PR gets reviewed.